+++ .gitignore
@@ -33,3 +33,6 @@ reflex
 *.jpg
 *.jpeg
 *.gif
+
+
+*.snap
