+++ Cargo.toml
@@ -15,6 +15,7 @@ clap = { version = "4.5", features = ["derive"] }
 
 # Error handling
 anyhow = "1.0"
+thiserror = "1.0"
 
 # Serialization
 serde = { version = "1.0", features = ["derive"] }
