+++ src/lib.rs
@@ -28,6 +28,7 @@
 pub mod ast_query;
 pub mod background_indexer;
 pub mod cache;
+pub mod errors;
 pub mod cli;
 pub mod context;
 pub mod content_store;
