+++ src/models.rs
@@ -323,6 +323,9 @@ pub struct IndexConfig {
     pub parallel_threads: usize,
     /// Query timeout in seconds (0 = no timeout)
     pub query_timeout_secs: u64,
+    /// Maximum entries per trigram posting list (0 = unlimited).
+    /// High-frequency trigrams are truncated at this threshold to bound query latency.
+    pub max_posting_list_entries: usize,
 }
 
 impl Default for IndexConfig {
@@ -335,6 +338,7 @@ impl Default for IndexConfig {
             max_file_size: 10 * 1024 * 1024, // 10 MB
             parallel_threads: 0, // 0 = auto (80% of available cores)
             query_timeout_secs: 30, // 30 seconds default timeout
+            max_posting_list_entries: 500_000, // cap at 500k to bound query latency
         }
     }
 }
