+++ src/semantic/providers/openai.rs
@@ -4,6 +4,7 @@ use super::LlmProvider;
 use anyhow::{Context, Result};
 use async_trait::async_trait;
 use serde_json::json;
+use std::time::Duration;
 
 /// OpenAI provider for GPT models
 pub struct OpenAiProvider {
@@ -14,9 +15,13 @@ pub struct OpenAiProvider {
 
 impl OpenAiProvider {
     /// Create a new OpenAI provider
-    pub fn new(api_key: String, model: Option<String>) -> Result<Self> {
+    pub fn new(api_key: String, model: Option<String>, timeout_secs: u64) -> Result<Self> {
+        let client = reqwest::Client::builder()
+            .timeout(Duration::from_secs(timeout_secs))
+            .build()
+            .context("Failed to build reqwest client")?;
         Ok(Self {
-            client: reqwest::Client::new(),
+            client,
             api_key,
             model: model.unwrap_or_else(|| "gpt-4o-mini".to_string()),
         })
@@ -172,14 +177,14 @@ mod tests {
 
     #[test]
     fn test_new_with_default_model() {
-        let provider = OpenAiProvider::new("test-key".to_string(), None).unwrap();
+        let provider = OpenAiProvider::new("test-key".to_string(), None, 300).unwrap();
         assert_eq!(provider.name(), "openai");
         assert_eq!(provider.model, "gpt-4o-mini");
     }
 
     #[test]
     fn test_new_with_custom_model() {
-        let provider = OpenAiProvider::new("test-key".to_string(), Some("gpt-4o".to_string())).unwrap();
+        let provider = OpenAiProvider::new("test-key".to_string(), Some("gpt-4o".to_string()), 300).unwrap();
         assert_eq!(provider.model, "gpt-4o");
     }
